Make gist selector a separate component
