+++ src/components/gist-select-widget/index.css
@@ -1,0 +1,83 @@
+.personal-gist {
+  min-width: 50%;
+  flex: 1;
+}
+
+.privacy-toggle {
+  display: flex;
+  align-items: center;
+  margin-bottom: 16px;
+}
+
+.privacy-toggle > input {
+  margin: 0 4px 0 0;
+}
+
+.gist-wrapper {
+  margin-right: 0.5em;
+  overflow-y: auto;
+  max-height: 47vh;
+}
+
+.load-gist {
+  flex: 1;
+}
+
+.loader-container {
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+  height: 30%;
+  font-size: 20px;
+}
+
+.text {
+  margin-left: 8px;
+  font-weight: bold;
+}
+
+.play-down {
+  color: var(--extra-dark-gray-color);
+}
+
+.personal-gist-description {
+  display: flex;
+  align-items: center;
+}
+
+.personal-gist-files {
+  display: flex;
+  flex-direction: column;
+}
+
+.gist-container {
+  margin-bottom: 1em;
+  font-size: 0.9em;
+  word-wrap: break-word;
+}
+
+.file {
+  display: flex;
+}
+
+.arrow {
+  margin: 4px;
+  width: 15px;
+  border-bottom: 1px solid var(--black-color);
+  border-left: 1px solid var(--black-color);
+}
+
+.filename {
+  font-family: monospace;
+  cursor: pointer;
+  margin-top: 4px;
+  color: var(--extra-dark-gray-color);
+}
+
+.filename:hover {
+  color: var(--black-color);
+}
+
+.loading {
+  opacity: 0.4;
+}
