+++ src/components/gist-select-widget/renderer.tsx
@@ -1,0 +1,195 @@
+import React from 'react';
+import {File, Lock} from 'react-feather';
+import ReactPaginate from 'react-paginate';
+import {mapStateToProps, mapDispatchToProps} from '.';
+import {BACKEND_URL, GistPrivacy} from '../../constants';
+import './index.css';
+
+interface State {
+  currentPage: number;
+  loaded: boolean;
+  personalGist: {
+    name: string;
+    isPublic: boolean;
+    title: string;
+    spec: {
+      name: string;
+      previewUrl: string;
+    }[];
+  }[];
+  pages: object;
+  loading: boolean;
+}
+
+type Props = ReturnType<typeof mapStateToProps> &
+  ReturnType<typeof mapDispatchToProps> & {
+    preview: (id?: string, file?: string, image?: string) => void;
+  };
+
+class GistSelectWidget extends React.Component<Props, State> {
+  constructor(props) {
+    super(props);
+    this.state = {
+      currentPage: 0,
+      loaded: false,
+      personalGist: [],
+      pages: {},
+      loading: true,
+    };
+  }
+  public componentDidMount() {
+    this.handlePageChange({selected: 0});
+  }
+  public componentDidUpdate(prevProps) {
+    if (this.props.isAuthenticated !== prevProps.isAuthenticated) {
+      this.setState(
+        {
+          currentPage: 0,
+          loading: true,
+        },
+        () => {
+          this.handlePageChange({selected: this.state.currentPage});
+        }
+      );
+    }
+    if (this.props.private !== prevProps.private) {
+      this.setState(
+        {
+          currentPage: 0,
+          loading: true,
+        },
+        () => {
+          this.handlePageChange({selected: this.state.currentPage});
+        }
+      );
+    }
+  }
+  public async handlePageChange(page) {
+    if (this.state.loading) {
+      this.setState(
+        {
+          loading: false,
+        },
+        async () => {
+          let response;
+          if (page.selected === 0) {
+            response = await fetch(`${BACKEND_URL}gists/user?cursor=init&privacy=${this.props.private}`, {
+              credentials: 'include',
+              method: 'get',
+            });
+          } else {
+            response = await fetch(
+              `${BACKEND_URL}gists/user?cursor=${this.state.pages[page.selected]}&privacy=${this.props.private}`,
+              {
+                credentials: 'include',
+                method: 'get',
+              }
+            );
+          }
+          const data = await response.json();
+          if (Array.isArray(data.data)) {
+            this.setState({
+              currentPage: page.selected,
+              loaded: true,
+              pages: page.selected === 0 ? data.cursors : this.state.pages,
+              loading: true,
+              personalGist: data.data,
+            });
+          } else {
+            this.props.receiveCurrentUser(data.isAuthenticated);
+          }
+        }
+      );
+    }
+  }
+  public render() {
+    const githubLink = (
+      /* eslint-disable-next-line react/jsx-no-target-blank */
+      <a href={`${BACKEND_URL}auth/github`} target="_blank">
+        Login with GitHub
+      </a>
+    );
+
+    return (
+      <React.Fragment>
+        <div className="personal-gist">
+          <h3>Your gists</h3>
+          {this.props.isAuthenticated ? (
+            this.state.loaded ? (
+              <>
+                <div className="privacy-toggle">
+                  <input
+                    type="checkbox"
+                    name="privacy"
+                    id="privacy"
+                    checked={this.props.private === GistPrivacy.ALL}
+                    onChange={this.props.toggleGistPrivacy}
+                  />
+                  <label htmlFor="privacy">Show private gists</label>
+                </div>
+                {this.state.personalGist.length > 0 ? (
+                  <>
+                    {Object.keys(this.state.pages).length > 1 && (
+                      <ReactPaginate
+                        previousLabel={'<'}
+                        nextLabel={'>'}
+                        breakClassName={'break'}
+                        containerClassName={'pagination'}
+                        activeClassName={'active'}
+                        pageCount={Object.keys(this.state.pages).length}
+                        onPageChange={this.handlePageChange.bind(this)}
+                        forcePage={this.state.currentPage}
+                        marginPagesDisplayed={2}
+                        pageRangeDisplayed={2}
+                      />
+                    )}
+                    <div className={`gist-wrapper ${!this.state.loading && 'loading'}`}>
+                      {this.state.personalGist.map((gist) => (
+                        <div key={gist.name} className="gist-container">
+                          <div className="personal-gist-description">
+                            {gist.isPublic ? (
+                              <File width="14" height="14" />
+                            ) : (
+                              <Lock width="14" height="14" fill="#FDD300" />
+                            )}
+                            <span className={`text ${gist.title ? '' : 'play-down'}`}>
+                              {gist.title ? gist.title : 'No description provided'}
+                            </span>
+                          </div>
+                          <div className="personal-gist-files">
+                            {gist.spec.map((spec, index) => (
+                              <div key={index} className="file">
+                                <div className="arrow"></div>
+                                <div
+                                  className="filename"
+                                  key={spec.name}
+                                  onClick={() => this.props.preview(gist.name, spec.name, spec.previewUrl)}
+                                >
+                                  {spec.name}
+                                </div>
+                              </div>
+                            ))}
+                          </div>
+                        </div>
+                      ))}
+                    </div>
+                  </>
+                ) : (
+                  <>You have no Vega or Vega-Lite compatible gists.</>
+                )}
+              </>
+            ) : (
+              <div className="loader-container">
+                <span>Loading your GISTS...</span>
+              </div>
+            )
+          ) : (
+            <span>{githubLink} to see all of your personal gist.</span>
+          )}
+        </div>
+      </React.Fragment>
+    );
+  }
+}
+
+export default GistSelectWidget;
