+++ src/components/header/gist-modal/index.css
@@ -3,78 +3,6 @@
   flex-wrap: wrap;
 }
 
-.personal-gist {
-  min-width: 50%;
-  flex: 1;
-}
-
-.load-gist {
-  flex: 1;
-}
-
-.privacy-toggle {
-  display: flex;
-  align-items: center;
-  margin-bottom: 16px;
-}
-
-.privacy-toggle > input {
-  margin: 0 4px 0 0;
-}
-
-.gist-wrapper {
-  margin-right: 0.5em;
-  overflow-y: auto;
-  max-height: 47vh;
-}
-
-.gist-container {
-  margin-bottom: 1em;
-  font-size: 0.9em;
-  word-wrap: break-word;
-}
-
-.personal-gist-description {
-  display: flex;
-  align-items: center;
-}
-
-.text {
-  margin-left: 8px;
-  font-weight: bold;
-}
-
-.play-down {
-  color: var(--extra-dark-gray-color);
-}
-
-.personal-gist-files {
-  display: flex;
-  flex-direction: column;
-}
-
-.file {
-  display: flex;
-}
-
-.arrow {
-  margin: 4px;
-  width: 15px;
-  border-bottom: 1px solid var(--black-color);
-  border-left: 1px solid var(--black-color);
-}
-
-.filename {
-  font-family: monospace;
-  cursor: pointer;
-  margin-top: 4px;
-  color: var(--extra-dark-gray-color);
-}
-
-.filename:hover {
-  color: var(--black-color);
-}
-
 .gist-text {
   margin-top: 30px;
   font-size: 12px;
@@ -109,14 +37,6 @@
 
 .gist-input-container input:checked + label {
   text-decoration: underline;
-}
-
-.loader-container {
-  display: flex;
-  justify-content: center;
-  align-items: flex-end;
-  height: 30%;
-  font-size: 20px;
 }
 
 .gist-input {
@@ -183,8 +103,4 @@
 
 .preview-error-fix {
   font-size: 12px;
-}
-
-.loading {
-  opacity: 0.4;
 }
